Update My_Minesweeper.py
- Managed to make it draw a 9 by 9 grid (previously drew a 8 by 8 grid)
- The buttons now display the number after clicking, just like the minesweeper game.
- Issues still pending: 
Populate is not populating the State_array properly, need to figure this out.
Also, the uncover function will never be reached due to the current structure. It always populates, and never goes to the actual game. The Movecount variable doesn't seem to be a good control for the games progress. Find alternative method.